feat(homepage): conversion fix-now batch from UX audit
- Homepage split into server page + HomePageClient: consumer-first
  metadata (title/OG/Twitter) overrides B2B default for message match
  with SEO/ad traffic; offerSlugFromParams moved with the client code
- Coverage-check form moved into the hero (above the fold, desktop and
  mobile); redundant scroll CTA removed, View packages kept as secondary
- Coverage-check failure: browser alert() replaced with inline error and
  WhatsApp fallback link pre-filled with the typed address
- Final CTA: mailto:sales@ replaced with tracked WhatsApp quote link via
  getWhatsAppLink() per contact-details rule
- Full audit report in docs/audits/2026-07-04_conversion-design-audit.md

Verified: jest offer-attribution test passes; type-check clean on all
touched files; dev-server render checked on desktop+mobile viewports.
Known: hydration warning in dev from Zoho cookie banner is pre-existing
and dev-only (reproduces on untouched /check-coverage; prod is clean).
